refactor: CheerTalk 모달 컴포넌트 수정
