+++ apps/spectator/components/cheertalk/Modal/CheerTalkModal.tsx
@@ -1,18 +1,16 @@
 import { Modal } from '@hcc/ui';
-import { useRef, useState } from 'react';
+import { useState } from 'react';
 
 import Banner from 'components/cheertalk/Modal/Banner';
-import CheerTalkForm from 'components/cheertalk/Modal/CheerTalkForm';
 import CheerTalkList from 'components/cheertalk/Modal/CheerTalkList';
 
 import AsyncBoundary from '@/components/common/AsyncBoundary';
 import Loader from '@/components/common/Loader';
 import useSocket from '@/hooks/useSocket';
-import GameCheerTalkFetcher from '@/queries/useGameCheerTalkById/Fetcher';
-import useSaveCheerTalkMutation from '@/queries/useSaveCheerTalkMutation/query';
 import { GameCheerTalkType } from '@/types/game';
 
 import * as styles from './CheerTalkModal.css';
+import ModalSection from './Section';
 
 interface CheerTalkModalProps {
   isOpen: boolean;
@@ -37,15 +35,6 @@ const CheerTalkModal = ({ isOpen, onClose, gameId }: CheerTalkModalProps) => {
 
   connect();
 
-  const { mutate } = useSaveCheerTalkMutation();
-
-  const scrollRef = useRef(null);
-  const scrollToBottom = () => {
-    if (!scrollRef.current) return;
-
-    (scrollRef.current as HTMLDivElement).scrollIntoView();
-  };
-
   return (
     <Modal isOpen={isOpen} onClose={onClose}>
       <div className={styles.wrapper}>
@@ -65,27 +54,7 @@ const CheerTalkModal = ({ isOpen, onClose, gameId }: CheerTalkModalProps) => {
           errorFallback={props => <CheerTalkList.ErrorFallback {...props} />}
           loadingFallback={<Loader />}
         >
-          <GameCheerTalkFetcher gameId={gameId}>
-            {({ gameTalkList, gameTeams, ...data }) => (
-              <div className={styles.cheerTalkListContainer}>
-                <ul className={styles.cheerTalkList}>
-                  <CheerTalkList
-                    cheerTalkList={gameTalkList.pages.flat()}
-                    scrollToBottom={scrollToBottom}
-                    {...data}
-                  />
-                  <CheerTalkList.SocketList cheerTalkList={cheerTalks} />
-                  <li ref={scrollRef}></li>
-                </ul>
-                <CheerTalkForm
-                  gameTeams={gameTeams}
-                  gameId={gameId}
-                  mutate={mutate}
-                  scrollToBottom={scrollToBottom}
-                />
-              </div>
-            )}
-          </GameCheerTalkFetcher>
+          <ModalSection gameId={gameId} cheerTalks={cheerTalks} />
         </AsyncBoundary>
       </div>
     </Modal>
